requirements.txt
pandas>=1.5.0,<2.0.0
plotly>=5.1.0
folium>=0.12.1
pytest>=6.2.5
python-dotenv>=0.19.0
geopandas>=0.9.0  # For handling geospatial data
branca>=0.4.2     # Required by folium for choropleth maps
streamlit>=1.22.0
streamlit-folium>=0.15.0
pandasai>=2.0.0,<3.0.0  # For AI-powered data analysis, using v2+ API
openai>=0.27.0,<1.0.0  # Required by pandasai
numpy>=1.24.0,<1.26.0  # Compatible with pandas 1.5.x and Python 3.11
sqlalchemy>=2.0.0  # For database connectivity with pandas
duckdb-engine>=0.9.0  # SQLAlchemy dialect for DuckDB 